+++ JHotDraw/src/main/java/org/jhotdraw/draw/DefaultDrawing.java
@@ -125,38 +125,6 @@ public class DefaultDrawing
         return null;
     }
 
-    // ToDo [aj] Smell = Write Code Once, QuadTreeDrawing class
-    public Figure findFigureBehind(Point2D.Double p, Figure figure) {
-        boolean isBehind = false;
-        for (Figure f : getFiguresFrontToBack()) {
-            if (isBehind) {
-                if (f.isVisible() && f.contains(p)) {
-                    return f;
-                }
-            } else {
-                isBehind = figure == f;
-            }
-        }
-        return null;
-    }
-
-    // ToDo [aj] Smell = Write Code Once, QuadTreeDrawing class
-    public Figure findFigureBehind(Point2D.Double p, Collection<? extends Figure> children) {
-        int inFrontOf = children.size();
-        for (Figure f : getFiguresFrontToBack()) {
-            if (inFrontOf == 0) {
-                if (f.isVisible() && f.contains(p)) {
-                    return f;
-                }
-            } else {
-                if (children.contains(f)) {
-                    inFrontOf--;
-                }
-            }
-        }
-        return null;
-    }
-
     public Figure findFigureExcept(Point2D.Double p, Collection<? extends Figure> ignore) {
         for (Figure f : getFiguresFrontToBack()) {
             if (!ignore.contains(f) && f.isVisible() && f.contains(p)) {
